Add a help panel as per mockup
Help panel implemented as another popover. Close icon styled using css. This
avoids issues with having to work out the precise right path, because we can
just work relative to the css.  When the close icon is clicked, close the help
popover. When the cheatsheet link is clicked, toggle the cheatsheet.

Conflicts:

	src/plugins/oer/math/lib/math-plugin.js
	src/plugins/oer/popover/lib/popover-plugin.coffee
	src/plugins/oer/popover/lib/popover-plugin.js