Add feature whereby code can be called after transaction
Because the MutationObserver is async and only processes the queue when
the local calling stack is empty, the entire tail-end of your code is
always recorded. This change allows you to defer the tail-end until after
the transaction is recorded. The idea was to use this for semantic block
activation, but in the end it wasn't useful for that.